ppocr_lite/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ppocr-lite"
version = "0.1.0"
description = "Lightweight PP-OCR runtime – onnxruntime + numpy + PIL, no OpenCV"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }

dependencies = [
    "numpy>=1.24",
    "Pillow>=9.0",
    "onnxruntime>=1.16",   # or onnxruntime-gpu
]

[project.optional-dependencies]
# scipy speeds up connected-component labelling; falls back to pure numpy without it
fast = ["scipy>=1.10"]

[tool.setuptools.packages.find]
where = ["."]
include = ["ppocr_lite*"]
